Renamed chart module and moved it to the Modules folder
