+++ wdip-fe/src/Modules/BarChart.tsx
